core/requirements_python.txt
numpy==1.21.5
sentence-transformers==2.2.0
huggingface-hub==0.10.1